old-tests/submission/Infraware/Forms/contents/Forms/form_method.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>Form_method</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
	<p><input type=hidden name="custname"></p>
	<p><input type=hidden name="custtel"></p>
	<p><input type=hidden name="custemail"></p>
 
	</form>
	<script>

		var form = document.getElementById("input_form");

		if (typeof(form.method) == "string") {
			test(function() {
				assert_equals(form.method, "post", "method attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("method attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
